[TESTKIT-1056] Improvements and overhaul of the toolbox (#1067)
* fix: cloud-init improvements for Windows

* fix: add reboot hook to launchpad for Windows

* fix: improved logging, timeout, for provisioner_clients

* fix: improve robustness and logging for ssh connections

* fix: improve logging, ease off on long sleeps

* fix: tentative improvement to multiline logging

* fix: breakpoints (commented out)

* fix: taint Windows nodes, to avoid MSR3 deployments landing there

* fix: sanitize command output, fix typo

* fix: change remote-exec command for Windows nodes

* fix: add Windows-specific support dump command

* fix: calculate max attempts based off max time / sleep time

* fix: add more logging to System()

* fix: add more logging, try/except, to Ucp()

* fix: add try/except block to SshClient()

* fix: cut WS reboot wait time to 0 secs

* fix: commenting out Windows taint (for now)

* fix: escape double-backslashes by doubling up

* fix: rename log archive tar+gzip file from .zip to .tar.gz

* fix: add containerd to Windows path

* fix: remove Windows reboot from launchpad template

* fix: improvements to Windows support dump + related logging

* fix: pass Linux/Windows check to sftp code

* fix: improve robustness of support dump, sftp handling

* fix: improve pod check loop/logging

* fix: add logging to sftp code

* fix: add Linux/Windows selectability to sftp code

* fix: improve Ucp() and _get_support_dump() logging

* fix: tweak ssh log msg

* fix: sftp relative paths, old code typo

* fix: make Windows dump command/output more legible

* chore: remove my temp commented code, breakpoint()'s

* chore: stylistic improvements to satisfy pycodestyle

* chore: fix pylint issues

* chore: fix a minor pylint issue

* fix: uncomment Windows taint method

+++ modules/templates/user_data_windows.tpl
@@ -2,8 +2,8 @@
 $admin = [adsi]("WinNT://./administrator, user")
 $admin.psbase.invoke("SetPassword", "${win_admin_password}")
 
-# Setting path to docker in advance so that docker commands resolve without a reboot to update env
-$updatedPath = $ENV:Path.TrimEnd(';') + ";$env:ProgramFiles\Docker"
+# Setting path to docker + containerd in advance so that docker and containerd commands resolve without a reboot to update env
+$updatedPath = $ENV:Path.TrimEnd(';') + ";$env:ProgramFiles\Docker;$env:ProgramFiles\containerd"
 [Environment]::SetEnvironmentVariable('Path', $updatedPath, 'Machine')
 
 # The Windows instance may not have the Containers feature (a requirement for MCR) installed
@@ -150,7 +150,34 @@ if ($rebootNeededForContainersFeature) {
     else {
         Write-Warning "Unable to install Windows Containers feature.  Exit code [$($feature.ExitCode)]."
     }
-    Write-Output "Restarting computer to complete initialization process."
-    Restart-Computer -Force
 }
+
+# Change hostname to conform to EC2 instance naming
+[bool] $rebootNeededForHostname = $false
+$metadataUrl = "http://169.254.169.254/latest/meta-data/hostname"
+Try {
+    $response = Invoke-Webrequest -Uri $metadataUrl -UseBasicParsing -ErrorAction Stop
+    if ($response.StatusCode -eq 200) {
+        if ((& hostname) -ne $response.Content) {
+            Write-Output "Changing hostname from [$(& hostname)] to [$($response.Content)]" 
+            Set-ItemProperty 'HKLM:\system\currentcontrolset\services\tcpip\parameters' -Name 'NV Hostname' -Value $response.Content
+            $rebootNeededForHostname = $true   
+        }
+        else {
+            Write-Output "Hostname [$(& hostname)] follows the expected naming convention and will not be changed." 
+        }
+    }
+    else {
+        Write-Warning "Received unexpected response code [$($response.StatusCode)] from EC2 instance metadata URL [$metadataUrl].  Unable to configure rename host."
+    }
+}
+Catch {
+    Write-Warning "Exception accessing URL [$metadataUrl]. Reason: $($_.Exception.Message).  Unable to rename host."
+}
+
+if (($rebootNeededForContainersFeature) -or ($rebootNeededForHostname)) {
+    Write-Output "Restarting computer to complete initialization process."
+    Restart-Computer -Force   
+}
+
 </powershell>
